requirements.txt
fastapi
python-dotenv
python-jose[cryptography]
pydantic
redis
celery
pydub
cloudinary
elevenlabs
requests
uvicorn
python-multipart
gevent
httpx
pyyaml==6.0
boto3
tenacity
